+++ viking.js
@@ -1165,17 +1165,17 @@ Viking.Coercions.Number = {
     load: function(value) {
         if (typeof value === 'string') {
             value = value.replace(/\,/g, '');
+			
+			if (value.trim() === '') {
+				return null;
+			}
         }
 
         return Number(value);
     },
 
     dump: function(value) {
-        if (typeof value === 'string') {
-            value = value.replace(/\,/g, '');
-        }
-
-        return Number(value);
+		return value;
     }
 };
 Viking.Coercions.String = {